Moved default options of AxisResizer.
