+++ ts/Extensions/DragPanes/AxisResizerOptions.d.ts
